+++ CHANGELOG.md
@@ -4,18 +4,24 @@
 
 ### Breaking changes
 
-- None
+- MeasurementControl -  raise a `ValueError` exception if there are no setpoints (!370)
 
 ### Merged branches and closed issues
 
+- Remove various old temp requirement pins (counterpart of quantify-scheduler!447) (!368)
 - Added support to SI_prefix_and_scale_factor for scaled units such as ns or GHz (!365)
 
 ## 0.6.3 (2022-08-05)
 
+### Breaking changes
+
+- MeasurementControl - Fix `print_progress` for 0 progress and do not raise when no setpoints (!363)
+
+### Merged branches and closed issues
+
 - Changes to InstrumentMonitor and PlotMonitor to fix errors due to racing conditions. (!358)
 - Can save metadata with `QuantifyExperiment` (!355)
 - Contribution guidelines updated, added: Versioning, Backward Compatibility and Deprecation Policy (!282)
-- MeasurementControl - Fix `print_progress` for 0 progress (!363)
 
 ## 0.6.2 (2022-06-30)