+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Remove various old temp requirement pins (counterpart of quantify-scheduler!447) (!368)
 - Added support to SI_prefix_and_scale_factor for scaled units such as ns or GHz (!365, !373)
+- InstrumentMonitor - Restore `initial_value` of `update_interval` param (removed in !324) (!375)
 
 ## 0.6.3 (2022-08-05)