+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Remove various old temp requirement pins (counterpart of quantify-scheduler!447) (!368)
 - Added support to SI_prefix_and_scale_factor for scaled units such as ns or GHz (!365, !373)
 - InstrumentMonitor - Restore `initial_value` of `update_interval` param (removed in !324) (!375)
+- Analysis - Use `long_name` instead of `name` XArray attribute as default for axis labels and plot messages (!380)
 
 ## 0.6.3 (2022-08-05)