+++ CHANGELOG.md
@@ -1,8 +1,6 @@
 # Changelog
 
-## Unreleased
-
-* Add functions to load and save lmfit fit results from file (!286)
+## 0.6.4 (2022-09-21)
 
 ### Breaking changes
 
@@ -12,6 +10,7 @@
 
 ### Merged branches and closed issues
 
+- Add functions to load and save lmfit fit results from file (!286)
 - Remove various old temp requirement pins (counterpart of quantify-scheduler!447) (!368)
 - Added support to SI_prefix_and_scale_factor for scaled units such as ns or GHz (!365, !373)
 - InstrumentMonitor - Restore `initial_value` of `update_interval` param (removed in !324) (!375)